* commit '0f03e23405b2759a8e10b4ae68506f456ea7d90c':
  修改我的零钱界面显示问题

+++ app/src/main/java/com/dkhs/portfolio/bean/AccountInfoBean.java
@@ -1,10 +1,12 @@
 package com.dkhs.portfolio.bean;
 
+import java.math.BigDecimal;
+
 /**
  * Created by wuyongsen on 2015/10/14.
  */
 public class AccountInfoBean {
-    public double available;
+    public BigDecimal available;
     public double frozen_trade;
     public double frozen_withdraw;
     public double profit_loss;
